refactor: move csv
