+++ src/app/_components/selected-panel.tsx
@@ -3,9 +3,9 @@ import { useRef } from 'react';
 import { useSelector } from 'react-redux';
 
 import { cn } from '@/app/(server)/_lib/cn';
-import { csvDownloadHandler } from '@/app/_lib/csv';
 import { clearAll, toggle } from '@/app/_services/selected-characters';
 import { useAppDispatch, type RootState } from '@/app/_store/store';
+import { csvDownloadHandler } from '@/app/api/csv/csv';
 
 import { CharacterThumbnail } from './ui/character-thumbnail';
